+++ exampleConnectionService/build.gradle
