+++ exampleConnectionService/images/call_to_bob.png
