+++ exampleConnectionService/images/dial_client_bob.png
