+++ exampleConnectionService/proguard-rules.pro
