+++ exampleConnectionService/src/main/res/drawable/ic_call_black_24dp.xml
