+++ exampleConnectionService/src/main/res/drawable/ic_call_end_white_24dp.xml
