+++ exampleConnectionService/src/main/res/layout/dialog_call.xml
