+++ exampleConnectionService/src/main/res/mipmap-hdpi/ic_launcher.png
