+++ exampleConnectionService/src/main/res/values/colors.xml
