+++ exampleConnectionService/src/main/res/values/dimens.xml
