+++ exampleConnectionService/src/main/res/values/styles.xml
